Use line-buffer print for stdout/stderr for git_send_email
Instead of flushing all stdout/stderr at last, print per-line as before the
rework on encoding, so that we can see patches posted one by one.

It's very good to have such interaction especially when posting large
patchsets, because in that case the git-send-email can take quite a while
and we totally have no idea whether the cmd hanged or it's just sending in
progress.  So this patch will recover the old behavior of subprocess.call.

Signed-off-by: Peter Xu <peterx@redhat.com>